Run Portal, webservice tests
