+++ .github/workflows/dotnet-core.yml
@@ -70,18 +70,18 @@ jobs:
     - name: Build
       run: dotnet build --nologo --configuration ${{ env.BUILD_CONFIG }} --no-restore
 
-    - name: Testing Backend
-      run: dotnet test --nologo --configuration ${{ env.BUILD_CONFIG }} --no-restore --filter FunderMaps.IntegrationTests.Backend
-      env:
-        ASPNETCORE_ENVIRONMENT: Production
-        UseExternalServices: False
-        Jwt__Issuer: "FunderMapsTest"
-        Jwt__Audience: "FunderMapsTest"
-        Jwt__TokenValidity: 10
-        Jwt__SignatureKey: somelargetextstringusedinsignatures
-        Incident__ClientId: 42
-        Incident__Recipients__0: info@example.com
-        ConnectionStrings__FunderMapsConnection: "Server=localhost;Database=fundermaps;User Id=fundermaps_webapp"
+    # - name: Testing Backend
+    #   run: dotnet test --nologo --configuration ${{ env.BUILD_CONFIG }} --no-restore --filter FunderMaps.IntegrationTests.Backend
+    #   env:
+    #     ASPNETCORE_ENVIRONMENT: Production
+    #     UseExternalServices: False
+    #     Jwt__Issuer: "FunderMapsTest"
+    #     Jwt__Audience: "FunderMapsTest"
+    #     Jwt__TokenValidity: 10
+    #     Jwt__SignatureKey: somelargetextstringusedinsignatures
+    #     Incident__ClientId: 42
+    #     Incident__Recipients__0: info@example.com
+    #     ConnectionStrings__FunderMapsConnection: "Server=localhost;Database=fundermaps;User Id=fundermaps_webapp"
 
     - name: Testing Portal
       run: dotnet test --nologo --configuration ${{ env.BUILD_CONFIG }} --no-restore --filter FunderMaps.IntegrationTests.Portal
